refactor: Rename audio_integrity to audio_reencoder
I've renamed the module `audio_integrity.py` to `audio_reencoder.py` to better reflect its functionality, which is primarily re-encoding audio streams.

- I renamed `ts2mp4/audio_integrity.py` to `ts2mp4/audio_reencoder.py`.
- I renamed `tests/test_audio_integrity.py` to `tests/test_audio_reencoder.py`.
- I updated all relevant import statements and mock paths in the tests and application code to reflect the new module name.
